feat: add type on services
