+++ src/services/trackingService.ts
@@ -1,0 +1,28 @@
+import { GetExperimentsRequest } from '@entities/experiment/repository';
+import { SearchRunRequest } from '@entities/run/repository';
+import ExperimentRepository from '@repositories/experimentRepository';
+import RunRepository from '@repositories/runRepository';
+
+class TrackingService {
+  private readonly experimentRepository: ExperimentRepository;
+  private readonly runRepository: RunRepository;
+
+  constructor() {
+    this.experimentRepository = new ExperimentRepository();
+    this.runRepository = new RunRepository();
+  }
+
+  async getExperiments(request: GetExperimentsRequest) {
+    const experiments = await this.experimentRepository.getExperiments(request);
+
+    return experiments;
+  }
+
+  async searchRun(request: SearchRunRequest) {
+    const runs = await this.runRepository.searchRun(request);
+
+    return runs;
+  }
+}
+
+export const trackingService = new TrackingService();
